Ajuste sql

+++ SQLQuery1.sql
@@ -13,8 +13,8 @@ Id INT NOT NULL IDENTITY,
 Name VARCHAR(60) NOT NULL,
 Description VARCHAR(200) NOT NULL,
 CreatedAt Date NOT NULL,
-Price Decimal NOT NULL,
-Quantity Decimal NOT NULL,
+Price Float NOT NULL,
+Quantity Float NOT NULL,
 )
 GO
